+++ src/main/java/hexlet/code/service/TaskStatusServiceImpl.java
@@ -1,6 +1,8 @@
 package hexlet.code.service;
 
 import hexlet.code.dto.TaskStatusDto;
+import hexlet.code.exception.DataException;
+import hexlet.code.model.Task;
 import hexlet.code.model.TaskStatus;
 import hexlet.code.repository.TaskStatusRepository;
 import lombok.AllArgsConstructor;
@@ -40,7 +42,13 @@ public class TaskStatusServiceImpl implements TaskStatusService {
     }
 
     @Override
-    public void deleteStatus(Long id) {
-        statusRepository.deleteById(id);
+    public void deleteStatus(Long id) throws DataException {
+        TaskStatus status = statusRepository.findById(id).orElseThrow();
+        List<Task> listOfTask = status.getTasks();
+        if (listOfTask == null || listOfTask.isEmpty()) {
+            statusRepository.deleteById(id);
+        } else {
+            throw new DataException("The Task status is associated with the task! Cannot be deleted! ");
+        }
     }
 }
