+++ src/main/java/hexlet/code/service/UserServiceImp.java
@@ -1,17 +1,19 @@
 package hexlet.code.service;
 
 import hexlet.code.dto.UserDto;
+import hexlet.code.exception.DataException;
+import hexlet.code.model.Task;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -24,6 +26,10 @@ public class UserServiceImp implements UserService {
 
     @Override
     public User createUser(UserDto userDto) throws Exception {
+        final Optional<User> existentUser = findUserByEmail(userDto.getEmail());
+        if (existentUser.isPresent()) {
+            throw new DataException("The User exist");
+        }
         final User user = new User();
         user.setEmail(userDto.getEmail());
         user.setFirstName(userDto.getFirstName());
@@ -34,6 +40,10 @@ public class UserServiceImp implements UserService {
 
     @Override
     public User updateUser(final long id, final UserDto userDto) throws Exception {
+        final Optional<User> existentUser = findUserByEmail(userDto.getEmail());
+        if (existentUser.isPresent()) {
+            throw new DataException("The User exist");
+        }
         final User userToUpdate = userRepository.findById(id).get();
         userToUpdate.setEmail(userDto.getEmail());
         userToUpdate.setFirstName(userDto.getFirstName());
@@ -53,7 +63,17 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws DataException {
+        User user = userRepository.findById(id).orElseThrow();
+        List<Task> listOfTaskAsAuthor = user.getTasksAsAuthor();
+        List<Task> listOfTaskAsExecutor = user.getTasksAsExecutor();
+        if (listOfTaskAsExecutor == null || listOfTaskAsAuthor == null) {
+            userRepository.deleteById(id);
+            return;
+        }
+        if (!listOfTaskAsAuthor.isEmpty() || !listOfTaskAsExecutor.isEmpty()) {
+            throw new DataException("The User is associated with the task! Cannot be deleted! ");
+        }
         userRepository.deleteById(id);
     }
 
@@ -66,8 +86,7 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public User findUserByEmail(String email) {
-        return userRepository.findUserByEmailIgnoreCase(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Sign in failed. User not found!"));
+    public Optional<User> findUserByEmail(String email) {
+        return userRepository.findUserByEmailIgnoreCase(email);
     }
 }
